changed search styling && created external stylesheet
