+++ screens/loginscreen/mainpage.js
@@ -1,5 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import {KeyboardAvoidingView, View, Text, TextInput, Image, FlatList, StyleSheet, TouchableOpacity, Keyboard, Modal } from 'react-native';
+//use external stylesheet
+import styles from '../../styles/MainPageStyles'; 
 
 const MainPage = ({ navigation, route }) => {
   const [isLoading, setIsLoading] = useState(true);
@@ -82,7 +84,7 @@ const MainPage = ({ navigation, route }) => {
 
 
             <Image
-                source={require('../../assets/icon.png')} // Placeholder image for post
+                source={require('../../assets/placeholder.jpg')} // Placeholder image for post
                 style={styles.postImage}
             />
 
@@ -105,20 +107,11 @@ const MainPage = ({ navigation, route }) => {
             style={styles.writeTaskWrapper}
             keyboardVerticalOffset={Platform.OS === "ios" ? 50 : -20} // Adjust the offset as needed
         >
+
             <View style={styles.searchContainer}>
-                <View style={styles.searchBarContainer}>
-                <TextInput
-                    style={styles.searchInput}
-                    placeholder="Search for an item"
-                    value={searchedItem}
-                    onChangeText={(text) => setSearchedItem(text)}
-                />
-                </View>
-                
                 {/* PLACEHOLDER FOR ADD BUTTON */}
                 {/* The navigation.navigate part must be the same for the popup to work. 
                 The current placeholder works but is not stylized. */}
-
                 <TouchableOpacity style={styles.addButton}
                     onPress={() => {
                         //send information to the main (current) page to "reset" the pop up.
@@ -131,12 +124,23 @@ const MainPage = ({ navigation, route }) => {
                         //navigate to the AddPage (where the user will actually end up)
                         navigation.navigate('AddPage')
                     }}>
-                      <Text>+</Text>
+                    <Image source={require('../../assets/add.png')} style={styles.addIconStyle} />
                 </TouchableOpacity>
                 {/* END OF PLACEHOLDER */}
+                
+                {/* Search Bar */}
+                {/* <View style={styles.searchBarContainer}>
+
+                    <TextInput
+                        style={styles.searchInput}
+                        placeholder="Search for an item"
+                        value={searchedItem}
+                        onChangeText={(text) => setSearchedItem(text)}
+                    />
+                </View> */}
 
                 <TouchableOpacity style={styles.searchButton} onPress={handleSearch}>
-                    <Text style={styles.buttonText}>Search</Text>
+                    <Image source={require('../../assets/search.png')} style={styles.searchIconStyle} />
                 </TouchableOpacity>
 
                 {/* PostPopup */}
@@ -158,183 +162,14 @@ const MainPage = ({ navigation, route }) => {
                     </Modal>
                 </View>
                 {/* PostPopup End */}
-
+                
             </View>
+
+
         </KeyboardAvoidingView>
+
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  // Add your styles here to format the feed items, headers, images, etc.
-  // This is a simplified example, and you may need to customize it further.
-  container: {
-    flex: 1,
-    backgroundColor: '#EDE7E7',
-  },
-  
-  postContainer: {
-    marginBottom: 20,
-    backgroundColor: '#fff',
-    margin: 28,
-    borderRadius: 20,
-    shadowColor: '#A59D95',
-    shadowOffset: {width: 0, height: 8},
-    shadowOpacity: 0.2,
-    shadowRadius: 24,
-    elevation: 7, //android shadow
-  },
-  row: {
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-    justifyContent: 'space-between',
-    marginHorizontal: 25,
-    marginVertical: 16,
-    alignItems: 'baseline',
-  },
-  itemName: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    fontWeight: '900',
-    // padding: 16,
-    fontSize: 30,
-    color: '#2F2E41',
-  },
-  username: {
-    fontWeight: '900',
-    fontSize: 16,
-    color: '#888',
-  },
-  date: {
-    fontSize: 14,
-    fontWeight: '700',
-    color: '#888',
-  },
-  postImage: {
-    width: '100%',
-    height: 200,
-    borderRadius: 20,
-
-  },
-  userDate: {
-    
-  },
-  description: {
-    marginBottom: 8,
-    color: '#2F2E41',
-    fontSize: 15,
-    fontWeight: 'bold',
-    marginTop: -3,
-  },
-  comments: {
-    color: '#888',
-  },
-
-//   input: {
-//     paddingVertical: 0,
-//     paddingHorizontal: 15,
-//     backgroundColor: '#FFF',
-//     borderRadius: 30,
-//     borderColor: '#C0C0C0',
-//     borderWidth: 1,
-//     width: 150,
-//   },
-
-  writeTaskWrapper: {
-    flex: 1,
-    backgroundColor: '#EDE7E7',
-    position: 'absolute',
-    bottom: 0,
-    paddingTop: 10,
-    width: '100%',
-    //height: 100,
-    flexDirection: 'row',
-    //justifyContent: 'space-around',
-    alignItems: 'center',
-  },
-
-  addWrapper: {
-    width: 50,
-    //height: 50,
-    backgroundColor: '#FFF',
-    borderRadius: 60,
-    justifyContent: 'center',
-    alignItems: 'center',
-    borderColor: '#C0C0C0',
-    borderWidth: 1,
-  },
-
-  addText: {},
-  searchButtonText: {
-    color: '#C2A3A3',
-    fontWeight: 'bold',
-  },
-
-  searchContainer: {
-    flexDirection: 'row',
-    backgroundColor: '#f0f0f0', // Background color of the search container
-    borderRadius: 15,
-    marginHorizontal: 10,
-    marginBottom: 50,
-  },
-
-  searchBarContainer: {
-    flex: 1,
-  },
-
-  searchInput: {
-    padding: 10,
-  },
-  searchButton: {
-    backgroundColor: '#FFAF66', // Background color of the search button
-    paddingVertical: 10,
-    paddingHorizontal: 20,
-    borderTopRightRadius: 10,
-    borderBottomRightRadius: 10,
-  },
-  buttonText: {
-    color: '#342F2F',
-    fontWeight: 'bold',
-  },
-
-  addButton: {
-    backgroundColor: '#C2A3A3', // Background color of the add button
-    paddingVertical: 10,
-    paddingHorizontal: 20,
-    borderRadius: 5,
-  },
-
-  //for PostPopup
-    //container for everything, makes space for it on the screen.
-  popupContainer: {
-      order: 3,
-      //backgroundColor: 'red', //for testing, to see where it is and how big it is.
-  },
-  //the actual part the text goes in.
-  popup: {
-      marginTop: '100%', //0% = top, 100% = center, 200% = bottom 
-      borderRadius: 7.5,
-      paddingHorizontal: 20,
-      height: 60, //constant for now
-
-      alignSelf: 'center',
-      backgroundColor: '#F04564',
-      alignItems: 'center',
-  },
-  postPopupText: {
-      color: '#2F2E41',
-      fontSize: 20,
-  },
-  popupButton: {
-      backgroundColor: '#FAF2F2',
-      height: '50%',
-      paddingHorizontal: '5%',
-      borderRadius: 10,
-
-      alignItems: 'center',
-      justifyContent: 'center',
-  },
-  //End of styles for PostPopup
-});
 
 export default MainPage;
